Feature/app optimization (#49)
* Remove unused dependency

* Remove unused dependency

* Remove unused dependency

* Remove unused dependency

* Cleanup code
