+++ app/build.gradle.kts
@@ -111,44 +111,18 @@ android {
 
 dependencies {
     implementation(project(":core"))
-    implementation("com.google.firebase:firebase-firestore:24.10.0")
-    implementation("com.google.firebase:firebase-storage:20.3.0")
-    implementation("com.google.firebase:firebase-auth:22.3.0")
-    implementation("androidx.lifecycle:lifecycle-runtime-ktx:2.6.2")
-    implementation("androidx.activity:activity-compose:1.8.2")
-    implementation(platform("androidx.compose:compose-bom:2023.08.00"))
-    implementation("androidx.compose.ui:ui")
-    implementation("androidx.compose.ui:ui-graphics")
-    implementation("androidx.compose.ui:ui-tooling-preview")
-    implementation("androidx.compose.material3:material3")
-    androidTestImplementation(platform("androidx.compose:compose-bom:2023.08.00"))
-    androidTestImplementation("androidx.compose.ui:ui-test-junit4")
 
-    // UI
-    applyBasicUiDependencies()
-    applyPagingDependencies()
-    applyComposeUiDependencies()
-
+    // Functions
     applyBasicFunctionDependencies()
-
-
+    applyAppSpecificDependencies()
+    applyFirebaseDependencies()
     applyNetworkDependencies()
     applyCoroutinesDependencies()
     applyLifecycleDependencies()
     applyHiltDependencies()
     applyLocalDbDependencies()
-    applyGlideDependencies()
 
-    // Specific dependencies
-    debugImplementation("com.squareup.leakcanary:leakcanary-android:2.12")
-
-
-
-    // Firebase BoM
-    implementation(platform("com.google.firebase:firebase-bom:32.7.0"))
-    // TODO: Add the dependencies for Firebase products you want to use
-    implementation("com.google.firebase:firebase-analytics")
-    debugImplementation("androidx.compose.ui:ui-tooling")
-    debugImplementation("androidx.compose.ui:ui-test-manifest")
-
+    // UI
+    applyComposeUiDependencies()
+    applyPagingDependencies()
 }
