+++ app/src/main/java/dev/baharudin/tmdb_android/presentation/MainActivity.kt
@@ -3,10 +3,7 @@ package dev.baharudin.tmdb_android.presentation
 import android.os.Bundle
 import androidx.activity.compose.setContent
 import androidx.appcompat.app.AppCompatActivity
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.Surface
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import androidx.navigation.compose.rememberNavController
 import dagger.hilt.android.AndroidEntryPoint
 import dev.baharudin.tmdb_android.presentation.navigations.AppNavHost
