+++ app/src/main/java/dev/baharudin/tmdb_android/presentation/screens/about/AboutScreen.kt
@@ -27,29 +27,25 @@ import androidx.compose.runtime.compositionLocalOf
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import coil.compose.AsyncImage
 import coil.compose.SubcomposeAsyncImage
 import dev.baharudin.tmdb_android.R
 
-val githubLightIcon = compositionLocalOf { R.drawable.ic_github_black }
-val githubDarkIcon = compositionLocalOf { R.drawable.ic_github_white }
+val localGithubLightIcon = compositionLocalOf { R.drawable.ic_github_black }
+val localGithubDarkIcon = compositionLocalOf { R.drawable.ic_github_white }
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun AboutScreen() {
 
     val githubIcon = if (isSystemInDarkTheme()) {
-        githubDarkIcon
+        localGithubDarkIcon
     } else {
-        githubLightIcon
+        localGithubLightIcon
     }
-
-    val context = LocalContext.current
 
     val openUrlLauncher = rememberLauncherForActivityResult(ActivityResultContracts.StartActivityForResult()) { }
