+++ app/src/main/java/dev/baharudin/tmdb_android/presentation/screens/detail/MovieDetailScreen.kt
@@ -1,11 +1,10 @@
 package dev.baharudin.tmdb_android.presentation.screens.detail
 
+import android.util.Log
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Tab
@@ -18,10 +17,11 @@ import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
 import androidx.paging.compose.collectAsLazyPagingItems
+import dev.baharudin.tmdb_android.presentation.navigations.NavigationItem
+import dev.baharudin.tmdb_android.presentation.navigations.buildRoute
 import dev.baharudin.tmdb_android.presentation.screens.detail.components.MovieDetailHeader
 import dev.baharudin.tmdb_android.presentation.screens.detail.components.MovieDetailInfo
 import dev.baharudin.tmdb_android.presentation.screens.detail.components.MovieReview
@@ -35,14 +35,12 @@ fun MovieDetailScreen(
     val reviews = viewModel.reviews.collectAsLazyPagingItems()
     var selectedTabIndex by remember { mutableIntStateOf(0) }
 
-    Scaffold(
-        // Scaffold parameters...
-    ) { innerPadding ->
+    Scaffold { innerPadding ->
         Box(
             modifier = Modifier.padding(innerPadding)
         ) {
             movie?.data?.let { movie ->
-                Column() {
+                Column {
                     MovieDetailHeader(movie) { viewModel.toggleFavoriteButton() }
 
                     TabRow(
@@ -63,7 +61,15 @@ fun MovieDetailScreen(
                     Box(modifier = Modifier.fillMaxHeight()) {
                         when (selectedTabIndex) {
                             0 -> {
-                                MovieDetailInfo(movie = movie) { /*TODO: Handle this */}
+                                MovieDetailInfo(movie = movie) { genre ->
+                                    val route = NavigationItem.MovieList.route.buildRoute(
+                                        mapOf(
+                                            "genreId" to genre.id, "genreName" to genre.name
+                                        )
+                                    )
+                                    Log.d("MovieDetailScreen", "navigate to $route")
+                                    navController.navigate(route)
+                                }
                             }
 
                             1 -> {
