--- app/src/main/res/menu/home_appbar_menu.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item
-        android:id="@+id/btn_about"
-        android:enabled="true"
-        android:icon="@android:drawable/ic_dialog_info"
-        app:showAsAction="never"
-        android:visible="true"
-        android:title="@string/about" />
-</menu>
