--- app/src/main/res/navigation/main_navigation.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/main_navigation"
-    app:startDestination="@id/homeFragment">
-    <fragment
-        android:id="@+id/homeFragment"
-        android:name="dev.baharudin.tmdb_android.presentation.home.HomeFragment"
-        android:label="HomeFragment" >
-        <action
-            android:id="@+id/action_homeFragment_to_detailFragment"
-            app:destination="@id/detailFragment" />
-        <action
-            android:id="@+id/action_homeFragment_to_aboutFragment"
-            app:destination="@id/aboutFragment" />
-        <action
-            android:id="@+id/action_homeFragment_to_movieListFragment"
-            app:destination="@id/movieListFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/detailFragment"
-        android:name="dev.baharudin.tmdb_android.presentation.detail.MovieDetailFragment"
-        android:label="fragment_detail"
-        tools:layout="@layout/fragment_movie_detail" >
-        <argument
-            android:name="movie"
-            app:argType="dev.baharudin.tmdb_android.core.domain.entities.Movie" />
-    </fragment>
-    <fragment
-        android:id="@+id/aboutFragment"
-        android:name="dev.baharudin.tmdb_android.presentation.about.AboutFragment"
-        android:label="fragment_about"
-        tools:layout="@layout/fragment_about" />
-    <fragment
-        android:id="@+id/movieListFragment"
-        android:name="dev.baharudin.tmdb_android.presentation.movie_list.MovieListFragment"
-        android:label="fragment_movie_list"
-        tools:layout="@layout/fragment_movie_list" >
-        <argument
-            android:name="genre"
-            app:argType="dev.baharudin.tmdb_android.core.domain.entities.Genre" />
-        <action
-            android:id="@+id/action_movieListFragment_to_detailFragment"
-            app:destination="@id/detailFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/movieReviewFragment"
-        android:name="dev.baharudin.tmdb_android.presentation.detail.movie_review.MovieReviewFragment"
-        android:label="MovieReviewFragment" >
-        <argument
-            android:name="movie"
-            app:argType="dev.baharudin.tmdb_android.core.domain.entities.Movie" />
-    </fragment>
-    <fragment
-        android:id="@+id/movieInfoFragment"
-        android:name="dev.baharudin.tmdb_android.presentation.detail.movie_info.MovieInfoFragment"
-        android:label="fragment_movie_info"
-        tools:layout="@layout/fragment_movie_info" >
-        <argument
-            android:name="movie"
-            app:argType="dev.baharudin.tmdb_android.core.domain.entities.Movie" />
-    </fragment>
-</navigation>
