+++ app/src/main/res/values/themes.xml
@@ -6,14 +6,4 @@
     </style>
 
     <style name="Theme.TheMovieDB" parent="Base.Theme.TheMovieDB" />
-
-    <style name="roundedCornersImageView" parent="">
-        <item name="cornerFamily">rounded</item>
-        <item name="cornerSize">8dp</item>
-    </style>
-
-    <style name="circleImageView" parent="">
-        <item name="cornerFamily">rounded</item>
-        <item name="cornerSize">50%</item>
-    </style>
 </resources>
