+++ buildSrc/src/main/java/UIDependencies.kt
@@ -1,0 +1,25 @@
+import org.gradle.api.artifacts.dsl.DependencyHandler
+
+fun DependencyHandler.applyComposeUiDependencies() {
+    val composeBom = platform("androidx.compose:compose-bom:2023.10.01")
+    implementation(composeBom)
+    androidTestImplementation(composeBom)
+    implementation("androidx.compose.ui:ui-graphics")
+    // Material Design 3
+    implementation("com.google.android.material:material:1.11.0")
+    implementation("androidx.compose.material3:material3:1.2.0-beta01")
+    // Android Studio Preview support
+    implementation("androidx.compose.ui:ui-tooling-preview")
+    debugImplementation("androidx.compose.ui:ui-tooling")
+    // UI Tests
+    androidTestImplementation("androidx.compose.ui:ui-test-junit4")
+    debugImplementation("androidx.compose.ui:ui-test-manifest")
+
+    implementation("androidx.activity:activity-compose:1.8.2") // Integration with activities
+    implementation("androidx.compose.runtime:runtime-livedata") // Integration with LiveData
+
+    implementation("io.coil-kt:coil:2.5.0")
+    implementation("io.coil-kt:coil-compose:2.5.0")
+
+    implementation(Dependencies.NAVIGATION_COMPOSE)
+}
