+++ buildSrc/src/main/java/Versions.kt
@@ -2,14 +2,12 @@ object Versions {
     const val AGP = "8.2.1"
     const val KSP = "1.9.21-1.0.16"
     const val KOTLIN = "1.9.21"
-    const val DAGGER = "2.49"
     const val HILT = "2.48"
     const val ROOM = "2.6.1"
     const val RETROFIT = "2.9.0"
     const val OKHTTP = "4.11.0"
     const val COROUTINES = "1.7.1"
-    const val LIFECYCLE = "2.6.2"
-    const val GLIDE = "4.16.0"
+    const val LIFECYCLE = "2.7.0"
     const val PAGING = "3.2.1"
     const val NAVIGATION = "2.7.6"
     const val COMPOSE_COMPILER = "1.5.7"
