+++ core/build.gradle.kts
@@ -66,10 +66,7 @@ android {
 }
 
 dependencies {
-    implementation("com.google.firebase:firebase-firestore:24.10.0")
-    implementation("com.google.firebase:firebase-storage:20.3.0")
-    implementation("com.google.firebase:firebase-auth:22.3.0")
-    implementation("androidx.constraintlayout:constraintlayout:2.1.4")
+    // Function
     applyBasicFunctionDependencies()
     applyPagingDependencies()
     applyNetworkDependencies()
