+++ core/src/main/java/dev/baharudin/tmdb_android/core/data/models/remote/get_movie_detail/ProductionCompanies.kt
@@ -1,13 +1,10 @@
-package dev.baharudin.tmdb_android
+package dev.baharudin.tmdb_android.core.data.models.remote.get_movie_detail
 
 import com.google.gson.annotations.SerializedName
 
-
-data class ProductionCompanies (
-
-  @SerializedName("id"             ) var id            : Int?    = null,
-  @SerializedName("logo_path"      ) var logoPath      : String? = null,
-  @SerializedName("name"           ) var name          : String? = null,
-  @SerializedName("origin_country" ) var originCountry : String? = null
-
+data class ProductionCompanies(
+    @SerializedName("id") var id: Int? = null,
+    @SerializedName("logo_path") var logoPath: String? = null,
+    @SerializedName("name") var name: String? = null,
+    @SerializedName("origin_country") var originCountry: String? = null
 )
